Added unit tests for `promtsdb.match2filter`.
