+++ promtsdb/promtsdb_test.go
@@ -1,119 +1,84 @@
-// +build integration
+// +build unit
+
+/*
+Copyright 2018 Iguazio Systems Ltd.
+
+Licensed under the Apache License, Version 2.0 (the "License") with
+an addition restriction as set forth herein. You may not use this
+file except in compliance with the License. You may obtain a copy of
+the License at http://www.apache.org/licenses/LICENSE-2.0.
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+implied. See the License for the specific language governing
+permissions and limitations under the License.
+
+In addition, you may not use the software for any purposes that are
+illegal under applicable law, and the grant of the foregoing license
+under the Apache 2.0 license is conditioned upon your compliance with
+such restriction.
+*/
 
 package promtsdb
 
 import (
-	"fmt"
 	"github.com/prometheus/prometheus/pkg/labels"
-	"github.com/prometheus/prometheus/storage"
-	"github.com/v3io/v3io-tsdb/pkg/config"
-	"github.com/v3io/v3io-tsdb/pkg/partmgr"
-	"math/rand"
+	"github.com/stretchr/testify/suite"
 	"testing"
-	"time"
 )
 
-const basetime = 15222481971234
-
-func TestTsdbIntegration(t *testing.T) {
-	t.Skip("Needs to be refactored - Doesnt test anything")
-
-	d, h := partmgr.TimeToDHM(basetime)
-	fmt.Println("base=", d, h)
-	cfg, err := config.LoadConfig("../v3io.yaml")
-	if err != nil {
-		t.Fatal(err)
-	}
-	fmt.Println(cfg)
-
-	adapter, err := NewV3ioProm(cfg, nil, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	//adapter.partitionMngr.GetHead().NextPart(0)
-
-	appender, err := adapter.Appender()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	lset := labels.Labels{labels.Label{Name: "__name__", Value: "http_req"},
-		labels.Label{Name: "method", Value: "post"}}
-
-	err = DoAppend(lset, appender, 50, 120)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	//time.Sleep(time.Second * 5)
-	//return
-
-	qry, err := adapter.Querier(nil, basetime-0*3600*1000, basetime+5*3600*1000)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	match := labels.Matcher{Type: labels.MatchEqual, Name: "__name__", Value: "http_req"}
-	match2 := labels.Matcher{Type: labels.MatchEqual, Name: "Aggregator", Value: "count,avg,sum"}
-	//params := storage.SelectParams{Func: "count,avg,sum", Step: 1000 * 3600}
-	params := storage.SelectParams{Func: "", Step: 0}
-	set, err := qry.Select(&params, &match, &match2)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	lasth := 0
-	for set.Next() {
-		if set.Err() != nil {
-			t.Fatal(set.Err())
-		}
-
-		series := set.At()
-		fmt.Println("\nLables:", series.Labels())
-		iter := series.Iterator()
-		//iter.Seek(basetime-1*3600*1000)
-		for iter.Next() {
-
-			if iter.Err() != nil {
-				t.Fatal(iter.Err())
-			}
-
-			t, v := iter.At()
-			d, h := partmgr.TimeToDHM(t)
-			if h != lasth {
-				fmt.Println()
-			}
-			fmt.Printf("t=%d:%d,v=%.2f ", d, h, v)
-			lasth = h
-		}
-		fmt.Println()
-	}
-
+type testPromTsdbSuite struct {
+	suite.Suite
 }
 
-func DoAppend(lset labels.Labels, app storage.Appender, num, interval int) error {
-	//return nil
-	//time.Sleep(time.Second * 1)
-	curTime := int64(basetime)
+func (suite *testPromTsdbSuite) TestMatch2filterEmpty() {
 
-	ref, err := app.Add(lset, curTime, 2)
-	if err != nil {
-		return err
+	name, filter, aggr := match2filter(nil)
+
+	suite.Require().Equal("", name)
+	suite.Require().Equal("", filter)
+	suite.Require().Equal("", aggr)
+}
+
+func (suite *testPromTsdbSuite) TestMatch2filterEqual() {
+
+	matchers := []*labels.Matcher{
+		{Type: labels.MatchEqual, Name: "field", Value: "literal"},
 	}
+	name, filter, aggr := match2filter(matchers)
 
-	for i := 0; i <= num; i++ {
-		time.Sleep(time.Millisecond * 80)
-		curTime += int64(interval * 1000)
-		t := curTime + int64(rand.Intn(100)) - 50
-		_, h := partmgr.TimeToDHM(t)
-		v := rand.Float64()*10 + float64(h*100)
-		fmt.Printf("t-%d,v%3.2f ", t, v)
-		err = app.AddFast(lset, ref, t, v)
-		if err != nil {
-			return err
-		}
+	suite.Require().Equal("", name)
+	suite.Require().Equal("field=='literal'", filter)
+	suite.Require().Equal("", aggr)
+}
+
+func (suite *testPromTsdbSuite) TestMatch2filterMultiple() {
+
+	matchers := []*labels.Matcher{
+		{Type: labels.MatchEqual, Name: "field1", Value: "literal1"},
+		{Type: labels.MatchNotEqual, Name: "field2", Value: "literal2"},
 	}
+	name, filter, aggr := match2filter(matchers)
 
-	return nil
+	suite.Require().Equal("", name)
+	suite.Require().Equal("field1=='literal1' and field2!='literal2'", filter)
+	suite.Require().Equal("", aggr)
+}
+
+func (suite *testPromTsdbSuite) TestMatch2filterMultipleWithName() {
+
+	matchers := []*labels.Matcher{
+		{Type: labels.MatchEqual, Name: "__name__", Value: "literal1"},
+		{Type: labels.MatchNotEqual, Name: "field2", Value: "literal2"},
+	}
+	name, filter, aggr := match2filter(matchers)
+
+	suite.Require().Equal("literal1", name)
+	suite.Require().Equal("field2!='literal2'", filter)
+	suite.Require().Equal("", aggr)
+}
+
+func TestPromTsdbSuite(t *testing.T) {
+	suite.Run(t, new(testPromTsdbSuite))
 }
